Fix committing to screen when another commit is pending
Sometimes a page-flip request won't make it to the immediately following
vblank due to being issued too close to it. When this happens, we don't want
to tell the client to render a new frame. This is the worst case scenario for
input lag (= a little more than the inverse of the refresh rate).